Keep count for each version
This means that we do not have to give the start number for every
question in map.yaml. It's a bit more fragile because just
counting will not always work -- q2 happens to contain an example
of this, where we have to retain the start value for a particular
version.