WIP
Some work in progress.

Hacking around with the first 9 questions,
trying to get this code to approximately work for all of them.

Note that I currently skip q6, and have also messed about with
its YAML definition.

Note that the way I work out question type does not make sense
(see TODO in code).

Over on the main branch, I might want to force some datatypes in
the CSV files that I'm reading -- want to make sure of consistent
datatypes in the index columns, or anything that might be involved
in a join or other lookup. The issue here is that I read the CSV
files separately -- so if one of them has an all-int column, where
another has a mixed-type column, those columns will have different
datatypes. If the same number is represented as a different type
across different dataframes then we are going to have a problem.

Recall that I have an almost-golden output example in
AAA_IGNORE/old_output. It is missing the intermediate/ dir and
q1_undecodable.